requirements.txt
fastapi
uvicorn
pillow
rembg
onnxruntime
mediapipe
opencv-python